Put asset functions into a package for re-use
Given that I've found myself copying these out into other projects, it
seems to make sense to put them into their own package for re-use.
